fix: improve phase precision and consistency across modulation DSPs
- Promote float phase accumulators to double in AM, PM, ASK, and CW
  sources to prevent long-run phase drift from mixed-precision
  arithmetic (double increment stored into float accumulator)
- Fix phase wrap edge case: change > to >= so phase landing exactly
  on 2π wraps correctly instead of accumulating past the boundary
- Deduplicate FmSource::render_block() duration logic by using shared
  compute_block_size() from SourceBase, matching all other sources
- Simplify FSK bandwidth formula removing dead (M-1)/(M-1) expression

+++ core/dsp/include/archerfish/dsp/am_source.hpp
@@ -22,7 +22,7 @@ private:
     double mod_freq_hz_{1000.0};
     double mod_depth_{0.5};
 
-    float carrier_phase_{0.0f};
+    double carrier_phase_{0.0};
 };
 
 } // namespace archerfish::dsp
